creación de componentes
