--- docs/asset-manifest.json
@@ -1,6 +1,0 @@
-{
-  "main.css": "static/css/main.3dc378ab.css",
-  "main.css.map": "static/css/main.3dc378ab.css.map",
-  "main.js": "static/js/main.d86ea3c9.js",
-  "main.js.map": "static/js/main.d86ea3c9.js.map"
-}
