--- docs/data/language.json
@@ -1,4 +1,0 @@
-{
-    "HTML": 5867,
-    "JavaScript": 2909
-  }
