--- docs/static/css/main.3dc378ab.css
@@ -1,2 +1,0 @@
-body{margin:0;padding:0;font-family:sans-serif}.App-logo{-webkit-animation:App-logo-spin infinite 20s linear;animation:App-logo-spin infinite 20s linear;height:80px}.App-header{background-color:#222;height:50px;padding:20px;color:#fff}.App-title{font-size:1.5em}.App-body{padding-top:20px}.App-intro{font-size:large}.usrprofile{border:1px solid #ccc;width:60%;height:100px;margin-bottom:10px}.usrprofile:hover{background-color:#ccc;cursor:pointer}.usrprofile>img{clear:both;margin-right:10px}.box{display:inline-block;vertical-align:top;margin:2px 5px}.contain{border:1px solid #c40404;width:60%}@-webkit-keyframes App-logo-spin{0%{-webkit-transform:rotate(0deg);transform:rotate(0deg)}to{-webkit-transform:rotate(1turn);transform:rotate(1turn)}}@keyframes App-logo-spin{0%{-webkit-transform:rotate(0deg);transform:rotate(0deg)}to{-webkit-transform:rotate(1turn);transform:rotate(1turn)}}.skill{margin:15px;padding:10px;text-align:center}.skill>div{height:50px;border:1px solid #ccc;margin:5px 10px;padding:2px;border-radius:1px 15px;display:grid;-ms-flex-line-pack:center;align-content:center;cursor:pointer}.skill>div:hover{background-color:#c0bfbf;border:1px solid #1ea0d3;font-weight:700}.show{display:block}.hide{display:none}
-/*# sourceMappingURL=main.3dc378ab.css.map*/
